+++ src/libv/ui/property/orientation.hpp.in.lua
@@ -1,6 +1,6 @@
 -- Project: libv.ui, File: src/libv/ui/property/orientation.hpp.in.lua, Author: Császár Mátyás [Vader]
 
-enum("Orientation", "int32_t")
+enum("Orientation", "int8_t")
 namespace("libv::ui")
 
 value("up",    "up")
